17545: Updates performance tests and removes accuracy tests (#16)
We are removing the accuracy tests since we have more rigorous accuracy
testing outside of this repository. Additionally, we are moving the
former MNIST accuracy test to the performance test directory and making
the adjustments to treat it as such.

All performance tests also have flags that indicate if the Trainee will
optimize, if the Trainee will use deviations, and if the Trainee will
use case weights.