17699: Adds a check to #assert_same to fail on incorrect usage (#31)
Previously, using #assert_same without passing the correct params would
lead to failing code appearing as if it passed. Here we add a check to
ensure an expected value is given. If it isn't given the test fails.

We also remove an inappropriate test, as it tests the unit test
framework itself, rather than the Engine.