19155: Adds checks for untrained features (#117)
Many methods returned '.nas' if called with parameters specifying
features that do not exist in the Trainee. This PR adds checks to some
of these methods that validate the context/action features parameters.

In the case an untrained feature is specified, the function returns
early specifying an error that details the exact parameter containing
the untrained feature.

I've chosen to add this method in a new Trainee module named
'validation'. This may seem premature, but I know we have future plans
to add increased validation within howso-engine, so this doesn't seem
inappropriate to me.