19745: Fixes an issue when forecasting with 'max_series_lengths' and time-series features with deltas (#146)
Fixes an issue that arose when users with time-series features using
deltas would call `react_series` with the `continue_series` flag and the
`max_series_lengths` parameter specified.

In this situation, the returned series data was often of less rows than
`max_series_lengths` would specify because the amount of rows pulled
into the generation to account for lags in the series being continued
counted towards the series length. This caused generation to terminate
early and the amount of generated cases to typically be the
max_series_length subtracted by the number of lags.

This change fixes the issue by increasing the internal maximum series
length by the amount of rows brought in to begin the forecast.