20608: Fixes reduce data removing all cases occasionally (#205)
Fixes `reduce_data` for datasets where most cases have an
`influence_weight_entropy` of 0. Previously, if too many cases had an
`influence_weight_entropy` of 0 when `reduce_data` was called the
threshold percentile had a chance of being equal to 0 which would remove
all of the cases in the model.

Also does parameter cleanup and moves the query that is shared by data
reduction and auto-ablation into its own method, to ensure unity.