21398: Updates the Amalgam unit test framework for ease of use (#279)
Changes the unit test framework in the following ways:

- Adds strict checks for the `exp` variable for assert methods that use
it. Calling these methods without setting the a non-null value for `exp`
will now trigger a failure
- Updated the value printing logic to always print values from a failed
assert if `#print_values_on_fail` is `(true)` which is on by default and
switches the previous `#verbose_on_fail` which caused all assert values
to be printed after a single failure to now be `(false)` by default.
- Parametrizes the sorting logic in `#assert_approximate`
- Some minor style/comment changes

Also updates the type schema and a unit test as "familiarity_conviction"
is not actually a functional react detail.