21479: Adds a warning to analyze when users do targeted analysis on a time-series Trainee with no previous targetless hyperparameters (#281)
Adds a warning to analyze that triggers when users call a targeted
analysis on a time-series Trainee that has no previously stored
targetless hyperparameters.

The motivation here is that training time-series data used to trigger a
targetless analysis internally. We strongly recommend targetless
hyperparameters when using the howso-engine for time-series data, but
with the removal of the internal analyze call in time-series flows,
users may only do a targeted analyze before calling things like
`#react_series`.

Here we add a new warning to analyze to try and warn users when they go
down this unrecommended path.