21587: Revises `#get_api` output by updating some methods idempotency and adding statisically_idempotent flags (#304)
Changed the logic for accumulating method annotations/labels into the
output of `#get_api`. Nothing is assumed to be True now, so any of the
flags ("read_only", "idempotent", "long_running", and now
"statistically_idempotent") must be manually specified as true to be
added to the object of information for a label.

If unspecified, the flag will not be present at all in the definition
for the label.