22320: Adds validation to Trainee names to allow only alphanumeric characters on hierarchy operations (#373)
This PR adds an extra layer of validation to hierarchy operations that
specify a Trainee name. These names must contain only alphanumeric
characters, so this PR adds logic to validate this for those labels.

This is done by adding new attribute, "pattern", to the type
specification for strings; and "min_length" and "max_length" to define
size limits for string. This allows the users/devs to specify a specific
regex pattern that string values must match as well as length
constraints for parameter validation

This wasn't enough though for all names that came through so there are
also a couple manual checks added for parameters where Trainee names
were given as indices within (assoc)'s.