22317: Updates `#update_trainee` to support updating contained subtrainees, MAJOR (#370)
Updates `#update_trainee` to account for subtrainees and upgrade them as
well in a recursive fashion allowing for the entire horde of Trainees to
be upgraded properly.

Also added a call to `#!ValidateParameters` within
`#get_num_training_cases` as it was missing there. This ensures that
this method is not called with invalid parameters (any parameters in the
case of this label).

This PR also removes the inactive parameter, `separate_files` from
`#upgrade_trainee`. Specifying this parameter after this change will
return an error that the `separate_files` parameter is not supported for
this label. Simply removing this parameter from the call should resolve
the issue.

---------

Co-authored-by: Cade Mack <24661281+cademack@users.noreply.github.com>