22488: Fixes an issue in `#react_series` where the first row of a series could only generate null feature values as all data was disallowed from the query (#392)
When doing series generation with a universal time feature, a situation
could often arise where the first row of a series would generate many
null values as the time value generated would be before that of any data
in the model. In this case, the internal queries to populate other
feature values would restrict the use of cases with time values larger
than that of the context, which would be all the training data in this
context.

Here lies a fix that puts a minimum bound on how restrictive this logic
can be. Specifically, the model will now always have access to at least
the K'th lowest time-value in the trained data.