22526: Refactors the `react_series` API, MAJOR (#405)
This set of changes presents a collection of major changes to the API
for both `#react_series` and `#single_react_series`. The collection of
changes is summarized in the following list of bullets.

- `leave_case_out` and 'case_indices' parameters are removed. These
parameters were not intuitive or useful in the context of
`#react_series`, they have been removed. This functionality is *not*
preserved.
- `action_values`, `context_features`, and `context_values` parameters
are removed. These were used to condition the stationary feature values
of a series. These were confusing to users and never used, so they are
removed. Conditioning a single row with one of these values in
`series_context_values` accomplishes the same goal.
- `leave_series_out`, `series_id_features`, and `series_id_values` have
been added. These parameters provide similar behavior to
`leave_case_out` and `case_indices` but operate at the series level.
Additionally, the `series_id_X` parameters provide users the ability to
select trained series for forecasting/reacting.
- `initial_features` and `initial_values` have been removed. These
served two purposes that were not completely obvious. The primary
function was to condition the first row of a generated series. This
functionality is maintained but should be achieved through specifying a
single row in `context_series_values`. The other function was to specify
the ID feature values of a trained series to forecast, this now should
be done with `series_id_features` and `series_id_values`.
- `continue_series_features` and `continue_series_values` have been
removed. To provide series data to forecast as these parameters once
supported, now pass the data in the same format to
`series_context_features` and `series_context_values` but additionally
set the `continue_series` flag to `(true)`.
- Renames 'use_regional_model_residuals` to `use_regional_residuals`