22759: Updates `react_aggregate` to use derivation logic for derived features and adds other logic to improve aggregate details when using time-series data (#424)
Updates `react_aggregate` to predict features using their derivation
logic rather than directly interpolating the values of influential
cases. For now this will be default behavior and will always apply to
features where it can, but it is planned to be parametrizable in the
future.

Additionally adds the time-series context filtering logic that has been
present in residuals to MDA and contributions.