22928: Refactors and fixes rebalance weights to work with ablation (#445)
changes:
- reduce auto re-analyze growth amount from e^2 to e since analyze is
much faster
- adds a new built in feature ".probability_mass" as
!internalLabelProbabilityMass for rebalance flows to store the
unweighted accumulated mass of a case. This value starts at 1.0 for all
cases and accumulates based on the amount of mass received from
ablated/removed cases.
- adds a new dataset attribute !cachedTotalMass which is the total
trained and accumulated mass of all cases ever trained (ablated or not)
into the dataset. This way if 10 cases were trained and 10 were ablated,
this !cachedTotalMass will be 20 and if the probability mass of the 10
remaining cases were added up it would also be 20
- adds a new dataset attribute !cachedRebalanceTotalMass which is the
total of all rebalanced and un-scaled case_weights. e.g., for a dataset
with a nominal rebalance feature with 2 classes, the unscaled rebalance
weight for each case is the reciprocal of its class count, once added up
will sum up to 2. additionally trained cases will increase this total
(until the next analyze)
- distance contributions being used for continuous rebalance features
are now scaled correctly by probability mass
- rebalanced case weights are scaled up by !cachedTotalMass /
!cachedRebalanceTotalMass so that the total of all trained cases' case
weights always adds up to !cachedTotalMass, (as it would if there was no
rebalancing), regardless whether rebalance features are nominal,
continuous or a mix.
- as cases are trained, they set their rebalance weights to an
approximately correct value in an attempt to keep the dataset rebalanced
(until the next analyze)
- for data reduction or ablation, as cases are ablated/removed and their
probability mass is distributed among the neighbors, the neighbor's case
weights are recomputed accordingly, that way when case weights are
recomputed on analyze, any ablation-accumulated or distributed mass is
not lost while maintaining rebalancing.