22983: Adds support for goal features and goal dependent features to react_aggregate, MINOR (#483)
Adds two parameters to `react_aggregate` which allow users to specify
goals that will bias the sampling of cases that will be used to compute
the desired metrics.

`goal_features_map` takes a map of feature names to goals (just like the
`react` equivalent). Instead of changing the logic of predictions
though, these specified goals instead alter the behavior of how the
cases are sampled to compute each metric. Cases are sampled in a way
that will still randomly select from the trained data, but will still
bias towards optimizing the defined goals for each sampled region of the
data.

`goal_dependent_features` takes a list of features and can be used to
specify features that will not be ignored while optimizing the goal in
the sampling process within each subregion of the data.