23473: Adds `use_all_features` flag to `react_series`, MINOR (#505)
Adds the `use_all_features` flag to `react_series`, which allows users
to disable the modelling of non-action features while series values are
generated when using react_series. This provides a strong boost to the
speed of forecasts and series generation when a subset of the trained
features are specified as action features.