23774: Fixes distance contributions and full similarity with all features when inactive features are present (#535)
sets feature weights for inactive features to 1 and provides them w the
minimal deviation to ensure that they are included in similarity
computations.
Reasoning is that if there are many features, with the majority having
individually same values, the cases are all mostly similar due to having
so many matching feature values, with only those few "active" features
for distinction. Thus when computing distance contributions or influence
weight entropy, all features should be used.