23819: Fixes a warning that can be emitted in forecasting regarding no trained series (#539)
A warning was written a bit incorrectly regarding what message was meant
to be sent to the user. Updated it to have one of two possible messages
depending on if there are truly no cases with the given series IDs or if
there are just no cases that have those IDs that are also before the
given starting time for the forecast.