23971: Checks that referenced entities exist in `#process_train_payload` (#561)
This is used when auto-ablation is on.  It's possible for one actor
to call compute_train_payload while another is executing a train-type
operation concurrently, which triggers automatic data reduction, which
causes child entities to be deleted.  When we do get to the "process"
step, double-check that all of the child entities we'd like to add
weights to are still there.