24057: Updates the math for the "feature_full_residual_convictions_for_case" detail to smooth out extremes (#570)
Rather than divide the regional residual by the case residual directly.
We now apply the LK metric to both residuals first, then divide. This
makes the extreme conviction values returned when either residual is
zero (or lesser than the deviation) much more reasonable to the user and
prevents random extremes that result when dealing with very predictable
regions of the data.