24507: Fixes autoanalyze during reduce data flow (#645)
when reduce_data() is kicked off during training w ablation, it may call
autoanalyze.
during analyze, it recomputes influenceWeightEntropy on a sample of
cases and nulls it out for the rest for performance, which breaks
reduce_data() since it needs them to be computed on all cases for proper
reduction.