24595: Update time-series internals regarding series termination/progress tracking, MAJOR (#663)
This is a significant rework to some of the internal feature-engineering
for time-series Trainees. An overview of the changes in this PR:

- `.series_progress` and `.series_progress_delta` are removed as
internal features.
- 'series_has_terminators' and 'stop_on_terminator' attributes have been
removed.
- Two new internal features are now derived automatically for TS data
- `.reverse_series_index`: zero based index for the series' timesteps in
reverse
- `.time_to_horizon`: A feature only used in universal time-feature
datasets that models how close the timesteps are to the "horizon" of the
dataset (the latest trained time-value).
- Series termination logic when no series-stop-maps are specified now
relies on the prediction of no following timesteps (rather than an
internally tracked series progress reaching 1.0)