24619: Adds support for a new time-series type: "covariate", MINOR (#667)
This PR adds support for a new time-series type, "covariate".

This joins the list of "rate" and "delta" as valid time-series types for
continuous time-series features. Rather than use a rate/delta and the
lag to determine the next feature value in a series, feature specified
as "covariate" have their next values in the series predicted directly
via interpolation (still using as many lags as specified as contexts).

These changes also make an adjustment to the dependent features logic
that automatically pulls deltas/rates into the dependent relationships
of their parents.